pyproject.toml
[project]
name = "thumbnail-generator"
version = "0.1.0"
description = "AI-powered YouTube thumbnail generator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "werkzeug>=3.0.0",
    "numpy>=1.24.0"
]

[project.scripts]
thumbnail = "thumbnail.main:main"
thumbnail-cli = "thumbnail.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["thumbnail"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
